# openapi.json
{
  "openapi": "3.0.0",
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development — بيئة التطوير"
    },
    {
      "url": "https://akfeek-backend.developteam.site",
      "description": "Production — بيئة الإنتاج"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token — أدخل التوكن بعد كلمة Bearer مثال: `Bearer eyJhbGci...`"
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "info": {
    "title": "🚗 Akfeek API — Full Reference",
    "version": "1.0.0",
    "description": "\n# Akfeek Platform — Full API Reference\n\n> **للمطورين الداخليين فقط** — هذه الوثيقة تحتوي على **جميع** الـ endpoints.\n> لو بتشتغل على جزء معين، استخدم الرابط المخصص ليك:\n\n| الجمهور | الرابط |\n|---|---|\n| 📱 تطبيق موبايل العميل | [`/api-docs/mobile`](/api-docs/mobile) |\n| 🔧 تطبيق موبايل الفني | [`/api-docs/technician`](/api-docs/technician) |\n| 🏪 بوابة الفيندور / الورشة | [`/api-docs/vendor`](/api-docs/vendor) |\n| ⚙️ لوحة تحكم الأدمن | [`/api-docs/admin`](/api-docs/admin) |\n\n---\n\n## Authentication / المصادقة\n\n```\nAuthorization: Bearer <your_jwt_token>\n```\n\n1. `POST /api/auth/login` — سجل دخول واحصل على التوكن\n2. انقر **Authorize** في الأعلى وأدخل: `Bearer <token>`\n      ",
    "contact": {
      "name": "Akfeek API Support",
      "email": "support@akfeek.com"
    }
  },
...
  "paths": {
...
    "/api/bookings": {
      "get": {
        "summary": "List all bookings (Admin)",
        "description": "Retrieve a paginated list of all bookings in the system.",
        "tags": [
          "📱 Customer | Bookings"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "Page number"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 10
            },
            "description": "Items per page"
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string"
            },
            "description": "Filter by status"
          }
        ],
        "responses": {
          "200": {
            "description": "List of bookings"
          }
        }
      },
      "post": {
        "summary": "Create new booking",
        "description": "Create a new booking with optional workshop and delivery method",
        "tags": [
          "📱 Customer | Bookings"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "vehicleId",
                  "scheduledDate"
                ],
                "properties": {
                  "customerId": {
                    "type": "string"
                  },
                  "vehicleId": {
                    "type": "string"
                  },
                  "scheduledDate": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "scheduledTime": {
                    "type": "string",
                    "example": "10:00"
                  },
                  "workshopId": {
                    "type": "string",
                    "description": "ID of certified workshop (optional)"
                  },
                  "deliveryMethod": {
                    "type": "string",
                    "enum": [
                      "FLATBED",
                      "SELF_DELIVERY"
                    ],
                    "description": "Required if workshopId is provided"
                  },
                  "serviceIds": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "productIds": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "notes": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Booking created successfully"
          },
          "400": {
            "description": "Invalid input or workshop not available"
          },
          "404": {
            "description": "Workshop not found"
          }
        }
      }
    },
...
    "/api/bookings/{id}": {
      "get": {
        "summary": "Get booking by ID",
        "description": "Retrieve details of a specific booking.",
        "tags": [
          "📱 Customer | Bookings"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Booking details"
          },
          "404": {
            "description": "Booking not found"
          }
        }
      }
    },
...
    "/api/bookings/{id}/status": {
      "patch": {
        "summary": "Update booking status",
        "description": "Update booking status. Auto-creates invoice when status is COMPLETED or DELIVERED.",
        "tags": [
          "⚙️ Admin | Bookings"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "status"
                ],
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": [
                      "PENDING",
                      "CONFIRMED",
                      "IN_PROGRESS",
                      "COMPLETED",
                      "DELIVERED",
                      "CANCELLED",
                      "TECHNICIAN_EN_ROUTE",
                      "ARRIVED",
                      "IN_SERVICE"
                    ]
                  },
                  "notes": {
                    "type": "string"
                  },
                  "technicianId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Status updated. Invoice auto-created if COMPLETED/DELIVERED."
          }
        }
      }
    },
...
    "/api/mobile-car-service/compatible-parts": {
      "get": {
        "summary": "Get compatible spare parts for a vehicle",
        "tags": [
          "📱 Customer | Mobile Car Service"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "vehicleId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of compatible parts"
          }
        }
      }
    },
    "/api/mobile-car-service/recommended-parts": {
      "get": {
        "summary": "Get recommended spare parts for a service",
        "tags": [
          "📱 Customer | Mobile Car Service"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "serviceId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of recommended parts"
          }
        }
      }
    },
    "/api/mobile-car-service/bookings": {
      "post": {
        "summary": "Create mobile car service booking",
        "tags": [
          "📱 Customer | Mobile Car Service"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "vehicleId",
                  "scheduledDate",
                  "serviceIds"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Booking created"
          }
        }
      }
    },
    "/api/mobile-car-service/bookings/{id}": {
      "get": {
        "summary": "Get mobile service booking details",
        "tags": [
          "📱 Customer | Mobile Car Service"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Booking details"
          }
        }
      }
    },
    "/api/mobile-car-service/bookings/{id}/status": {
      "patch": {
        "summary": "Update mobile service booking status",
        "tags": [
          "📱 Customer | Mobile Car Service"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "status"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Status updated"
          }
        }
      }
    },
...
    "/api/services": {
      "get": {
        "summary": "Get all services",
        "description": "Retrieve a list of all available services with filtering",
        "tags": [
          "📱 Customer | Services"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "category",
            "schema": {
              "type": "string",
              "enum": [
                "CLEANING",
                "MAINTENANCE",
                "REPAIR",
                "EMERGENCY",
                "INSPECTION",
                "CUSTOMIZATION",
                "COMPREHENSIVE_CARE"
              ]
            }
          },
          {
            "in": "query",
            "name": "type",
            "schema": {
              "type": "string",
              "enum": [
                "FIXED",
                "CATALOG",
                "EMERGENCY",
                "INSPECTION"
              ]
            }
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of services retrieved"
          }
        }
      },
      "post": {
        "summary": "Create new service (Admin)",
        "tags": [
          "⚙️ Admin | Services"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "type",
                  "category"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "nameAr": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "type": {
                    "type": "string"
                  },
                  "category": {
                    "type": "string"
                  },
                  "estimatedDuration": {
                    "type": "integer"
                  },
                  "pricing": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "vehicleSize": {
                          "type": "string"
                        },
                        "basePrice": {
                          "type": "number"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Service created"
          }
        }
      }
    },
...
    "/api/services/{id}": {
      "get": {
        "summary": "Get service details",
        "tags": [
          "📱 Customer | Services"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Service details retrieved"
          },
          "404": {
            "description": "Service not found"
          }
        }
      },
      "put": {
        "summary": "Update service (Admin)",
        "tags": [
          "⚙️ Admin | Services"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "isActive": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Service updated"
          }
        }
      },
      "delete": {
        "summary": "Delete/Deactivate service (Admin)",
        "tags": [
          "⚙️ Admin | Services"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Service deleted"
          }
        }
      }
    },
...
  }
}